requirements.txt
aiofiles==24.1.0
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.3.0
certifi==2025.8.3
cffi==1.17.1
click==8.2.1
cryptography==45.0.5
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iso8601==2.1.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
passlib==1.7.4
pyasn1==0.6.1
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pypika-tortoise==0.6.1
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
rich==14.1.0
rich-toolkit==0.14.9
rignore==0.6.4
rsa==4.9.1
sentry-sdk==2.34.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.47.2
tortoise-orm==0.25.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
